pyproject.toml
[tool.poetry]
name = "ai01"
version = "0.2.14"
description = "AI Agent SDK on Huddle01 dRTC Network"
authors = ["Om Gupta <omgupta0720@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "2.10.1"
ruff = ">=0.6.0,<=0.9.1"
asyncio = "^3.4.3"
pyee = "11.1.0"
av = "11.0.0"
numpy = "^2.1.3"
websockets = "^13.1"
huddle01 = "1.0.11"
uuid = "^1.30"
google-genai = "^0.5.0"
sounddevice = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
